chore(rich-text-editor): reflect current selection in text block selector
