+++ libs/components/src/lib/rich-text-editor/menubar/menubar.ts
@@ -1,4 +1,4 @@
-import { attr, observable } from '@microsoft/fast-element';
+import { attr } from '@microsoft/fast-element';
 import { VividElement } from '../../../shared/foundation/vivid-element/vivid-element';
 import type { RichTextEditor } from '../rich-text-editor';
 
@@ -6,24 +6,31 @@ export class MenuBar extends VividElement {
 	@attr({ attribute: 'menu-items' })
 	menuItems?: string;
 
-	@observable
-	textStylesState: Record<string, any> = {};
-
 	get #textEditorElement() {
 		return this.parentElement as RichTextEditor;
 	}
 
-	#updateTextStyleState: EventListener = ((event: Event) => {
-        this.textStylesState = (event as CustomEvent).detail;
-    }) as EventListener;
-	
+	#updateTextStyleState: EventListener = (() => {
+		this.dispatchEvent(
+			new CustomEvent('text-styles-changed', {
+				detail: this.#textEditorElement?.selectionStyles,
+			})
+		);
+	}) as EventListener;
+
 	override connectedCallback(): void {
 		super.connectedCallback();
-		this.#textEditorElement.addEventListener('selection-changed', this.#updateTextStyleState);
+		this.#textEditorElement.addEventListener(
+			'selection-changed',
+			this.#updateTextStyleState
+		);
 	}
 
 	override disconnectedCallback(): void {
 		super.disconnectedCallback();
-		this.#textEditorElement?.removeEventListener('selection-changed', this.#updateTextStyleState);
+		this.#textEditorElement?.removeEventListener(
+			'selection-changed',
+			this.#updateTextStyleState
+		);
 	}
 }
